% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathstuff.R
\name{addition}
\alias{addition}
\alias{multiplication}
\title{Add or multiply two numbers}
\usage{
addition(a, b)

multiplication(a, b)
}
\arguments{
\item{a}{a number}

\item{b}{a number}
}
\value{
\code{addition()}: the sum of \code{a} and \code{b}. \code{multiplication()}: the product of \code{a} and \code{b}.
}
\description{
Add or multiply two numbers
}
\examples{
addition(1, 2)
multiplication(4, 4)
}
